pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = "setuptools.build_meta"

[project]
name = "cosmodiff"
description = "Diffusion models for cosmology"
authors = [{name="Nicholas Kern"}, {name="Jiaming Pan"}]
classifiers = [
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "numpy",
    "torch",
    "diffusers",
    "accelerate",
    "tqdm",
    "pyyaml",
    "scipy",
    "h5py",
    "matplotlib",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "coverage",
    "ema-pytorch",
]
docs = [
    "cosmodiff[dev]",
    "furo",
    "nbsphinx",
    "sphinx>=1.8,<7.2",
    "myst-nb",
    "myst_parser",
]

[project.urls]
Repository = "https://github.com/nkern/cosmo_diffusion"

[tool.pytest.ini_options]
addopts = "--cov cosmodiff --cov-config=.coveragerc --cov-report xml:./coverage.xml"
testpaths = 'cosmodiff/tests'
filterwarnings = ["ignore::DeprecationWarning"]

[tool.setuptools]
script-files = ["scripts/cosmodiff_train.py"]

[tool.setuptools.package-data]
"cosmodiff.data" = ["*.npy", "*.yaml", "*.txt"]

